requirements.txt
streamlit
plotly
streamlit-option-menu
mysql-connector-python